backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coverage>=7.8.2",
    "django>=5",
    "django-cors-headers>=4.3.1",
    "django-debug-toolbar>=5.2.0",
    "django-filter>=25.1",
    "django-ratelimit>=4.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.2.1",
    "requests>=2.31.0",
    "openai>=1.3.7",
    "sentence-transformers",
    "faiss-cpu",
    "pdfplumber",
    "unidecode",
    "django-redis>=5.4.0",
    "gunicorn>=21.2.0",
    "python-dotenv>=1.1.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "google-generativeai>=0.8.5",
    "dj-database-url>=3.0.1",
]
